+++ src/components/index.ts
@@ -1,3 +1,4 @@
+export * from './base';
 export * from './delete';
 export * from './form';
 export * from './list';